pyproject.toml
[project]
name = "thor"
version = "0.1.0"
description = "A Python 3.14 micro web framework with secure cookies, sessions, auth, and lifespan management"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyjwt[crypto]>=2.11.0",
    "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/thor"]

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "httpx>=0.28.1",
    "mypy[mypyc]>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]
